(Test) fix filename of the golden references in tests and rename tpgGenerationEngine and programGenerationEngine with a lower case
