+++ gegelatilib/include/gegelati.h
@@ -44,8 +44,8 @@
 #define GEGELATI_H
 
 #ifdef CODE_GENERATION
-#include <code_gen/ProgramGenerationEngine.h>
-#include <code_gen/TpgGenerationEngine.h>
+#include <code_gen/programGenerationEngine.h>
+#include <code_gen/tpgGenerationEngine.h>
 #endif // CODE_GENERATION
 
 #include <data/array2DWrapper.h>
